Prisma/GraphQL: convert judge confidenceScore to decimal/float (0.00–1.00)
- Prisma: change adaptiveRubricJudgeRecord.confidenceScore from Int (1-5) to Decimal? @db.Decimal(5,2) and update comment to 0.00 - 1.00 scale
- GraphQL: update JudgeRecord.confidenceScore type from Int to Float

+++ src/graphql/type/TypeDefs.ts
@@ -98,7 +98,7 @@ export const typeDefs = `#graphql
     adaptiveRubricId: Int!
     accountId: String!
     result: Boolean!
-    confidenceScore: Int
+    confidenceScore: Float
     notes: String
     judgedAt: DateTime!
   }
